Project 2 Overview.
what we need to do in project 2.

### Overview

In this project, the student had to make use of Python to explore data related to bike share systems for three major cities in the United States — Chicago, New York City, and Washington. The student had to write code to (a) import the data and answer interesting questions about it by computing descriptive statistics, and (b) write a script that takes in raw input to create an interactive experience in the terminal to present these statistics.

### Project Submission

The developed CLI program allows the user to explore an US bikeshare system database and retrieve statistics information from the database. The user is able filter the information by city, month and weekday, in order to visualize statistics information related to a specific subset of data. The user is also able to chose to view raw data and to sort this data by columns, in ascending or descending order.

[Check My Project Submission](https://github.com/kadryjr/Udacity-programming-for-Data-Science-using-Python-Nanodegree/tree/master/Project%202).

#### Files Used

The required files for running this program are: 

* washington.csv
* new_york_city.csv
* chicago.csv

#### Requirements
This program was written in Python (version 3.7.1) and relies on the following libraries:

* pandas==0.23.4
* numpy==1.15.4